### 2018-04-21 #686 @leeper
@anhoej Thanks for your submission! Can you please fix the errors in your paper.md that are leading the PDF to not be generated. Looks like the issue is use of an svg image. Perhaps try PNG instead?
