### 2018-04-22 1cd7a74
Convert paper.svg to paper.png.
