### 2018-04-22 #686 @anhoej
Thanks! Try now.

Jacob

2018-04-21 21:41 GMT+02:00 Thomas J. Leeper <notifications@github.com>:

> @anhoej <https://github.com/anhoej> Thanks for your submission! Can you
> please fix the errors in your paper.md that are leading the PDF to not be
> generated. Looks like the issue is use of an svg image. Perhaps try PNG
> instead?
>
