### 2018-04-23 #686 @anhoej
Must the editor be on your list of editors? I don't see any with expertise
in statistical process controll. But maybe that's not important. Most
people with a high school degree should be able to get the idea after
reading the vignette. Or maybe subject matter knowledge is not at all
important if you know R.

What I'm trying to say is that I'm not sure how this works and how you pick
editors. But I will be glad to help if I can in finding somebody who can be
the editor on my paper.

Kind regards
Jacob

2018-04-23 0:22 GMT+02:00 Thomas J. Leeper <notifications@github.com>:

> Super. Mange tak! Do let me know if you have any suggested reviewers for
> this software.
>
