### 2018-04-23 #686 @leeper
@anhoej I'll serve as editor for this submission and will recruit one or more reviewers. Just wanted to see if you already had someone in mind as a possible reviewer. No problem that you don't!
