### 2018-04-23 #686 @leeper
@rmadillo Would you be interested in reviewing this submission for JOSS?
