### 2018-04-24 #699 @whedon
[REVIEW]: qicharts2: Quality Improvement Charts for R
**Submitting author:** @anhoej (<a href="http://orcid.org/0000-0002-7701-1774">Jacob Anhøj</a>)
**Repository:** <a href="https://github.com/anhoej/qicharts2" target ="_blank">https://github.com/anhoej/qicharts2</a>
**Version:** 0.4.0.9000

- [x] **Version:** Does the release version given match the GitHub release (0.4.0.9000)?
