### 2018-04-27 6405c48
Update paper.md
Fixed a few typos, made a few suggestions for phrasing.
