### 2018-05-07 #699 @Rmadillo
Hi @leeper,

I've finished the review and sent 2 pull requests to the master branch for paper.md and the vignette for small typos and clarifications.

I tested all code within the vignette and man page examples. I did not review the math in the code, but did do a spot check against manually-calculated values to see if they were the same (they were).

- I might like to have seen the `show.grid = T` in the first case instead of the third of the vignette.
- Does the `cagb` data set have/need a source? Other example dataframes have sources in their man pages, but this one does not.
- There was a co-author in the original `qicharts`--was that because they did the coding for the charts in `lattice` and this one is done in `ggplot2`?
- Is there a reason the `runvals` option was removed in this version? I liked that feature in `qicharts`.

Other than that,  no real criticisms--it's a really great package. My review recommendation is to accept.

Apologies that this took so long! Life has a funny way of intervening when you're trying to complete things. :D
