### 2018-05-08 #699 @anhoej
Thanks alot!

1. I find that the grid only makes sense when you need to identify the x-value of  individual data points, which is only the case in case 3.

2. The source of the cabg dataset is deliberately not revealed to avoid potential identification of individual patients from the data provided.

3. I had great help from my student assistant, Timo, with qichart. He, however left to finish his studies (you people, you know), and qicharts2 no longer needed the functionality he created.

4. Yes, runsvals = TRUE in qicharts printed detailed information from the runs analysis of single charts directly on the plot. However, since plots in qicharts2 may become very complicated with one and two dimensional facets and two or more chart parts, I decided not to try and implement this functionality. Instead, I made the summary function that reveals the same detailed information on each facet and each part of a qic plot. This information may be captured and presented, for example, in a figure caption or in any text or table accompanying the chart.

Thank you for a very pleasant submission review process. And thank you for the corrections and your suggestions to improve qicharts2.

Kind regards
Jacob Anhøj
