### 2018-05-09 #699 @Rmadillo
@leeper absolutely. I'd say it's _completely_ up to @anhoej, but it might be nice to have a "Source" section for `cabg` that says something like "Omitted for privacy concerns", mostly just so it doesn't look like an oversight, given that the other data sets do show a source section. Of course, I can also see reasons to leave it blank; if someone _knew_ it was real data, they could work to de-anonymize it, much like the NYC Taxi data issue. So either way is fine, depending on @anhoej's preferences. I consider this review/response complete, so feel free to close it.

@anhoej my apologies for not thinking through a few things a little more carefully; of course #4 would be a nightmare with the new charting options. And, ultimately, thanks for an incredibly useful package; it's what I've started recommending to my colleagues for their SPC uses.

Best wishes.
