### 2018-05-09 #699 @anhoej
Perfect! Thanks! Very good suggestion. I'll add "Omitted for privacy
concers" as data source for the cabg data set.

@leeper, I'm almost ready to release an update. Should I wait for the paper?

Kind regards
Jacob

2018-05-09 19:08 GMT+02:00 Rmadillo <notifications@github.com>:

> @leeper <https://github.com/leeper> absolutely. I'd say it's *completely*
> up to @anhoej <https://github.com/anhoej>, but it might be nice to have a
> "Source" section for cabg that says something like "Omitted for privacy
> concerns", mostly just so it doesn't look like an oversight, given that the
> other data sets do show a source section. Of course, I can also see reasons
> to leave it blank; if someone *knew* it was real data, they could work to
> de-anonymize it, much like the NYC Taxi data issue. So either way is fine,
> depending on @anhoej <https://github.com/anhoej>'s preferences. I
> consider this review/response complete, so feel free to close it.
>
> @anhoej <https://github.com/anhoej> my apologies for not thinking through
> a few things a little more carefully; of course #4 would be a nightmare
> with the new charting options. And, ultimately, thanks for an incredibly
> useful package; it's what I've started recommending to my colleagues for
> their SPC uses.
>
> Best wishes.
>
