### 2018-05-10 #699 @anhoej
Thanx. Will do. It might take a while. I'm waiting for R 3.5.0 for Linux to
be released before I build and release the new package on CRAN.

2018-05-09 20:22 GMT+02:00 Thomas J. Leeper <notifications@github.com>:

> @anhoej <https://github.com/anhoej> Okay, great. I think that's it then.
> Please make any final updates, issue a release, and complete an archive of
> the package through, e.g., figshare or zenodo. Please post the DOI for that
> archive once you have it and then we can complete the review process!
>
> Thank you, again, @Rmadillo <https://github.com/Rmadillo>!
>
