### 2018-05-10 #699 @leeper
@anhoej Sorry, just meant a [Github release](https://help.github.com/articles/creating-releases/). No need to submit to CRAN at this point.
