### 2018-05-10 #699 @anhoej
OK, thanks!

2018-05-10 11:13 GMT+02:00 Thomas J. Leeper <notifications@github.com>:

> @anhoej <https://github.com/anhoej> Sorry, just meant a Github release
> <https://help.github.com/articles/creating-releases/>. No need to submit
> to CRAN at this point.
>
