### 2018-05-10 #699 @anhoej
I'm new to this DOI thing. Is this correct?
https://github.com/anhoej/qicharts2

2018-05-10 11:14 GMT+02:00 Jacob Anhøj <jacob@anhoej.net>:

> OK, thanks!
>
> 2018-05-10 11:13 GMT+02:00 Thomas J. Leeper <notifications@github.com>:
>
>> @anhoej <https://github.com/anhoej> Sorry, just meant a Github release
>> <https://help.github.com/articles/creating-releases/>. No need to submit
>> to CRAN at this point.
>>
>> —
>> You are receiving this because you were mentioned.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/openjournals/joss-reviews/issues/699#issuecomment-387999474>,
>> or mute the thread
>> <https://github.com/notifications/unsubscribe-auth/AEQ_xMYFykX0ussfyiYOnZXNSfbKSP53ks5txASngaJpZM4ThLnU>
>> .
>>
>
>
>
> --
> Venlig hilsen
> Jacob Anhøj
>

--
Venlig hilsen
Jacob Anhøj
