### 2018-05-10 5d8b185
Add x.period argument.
