### 2018-05-11 9753326
Hyperlink DOIs against preferred resolver
